Changed futures api quantity parameter to int

+++ Kucoin.Net/Clients/FuturesApi/KucoinClientFuturesApiTrading.cs
@@ -35,7 +35,7 @@ namespace Kucoin.Net.Clients.FuturesApi
             OrderSide side,
             NewOrderType type,
             decimal leverage,
-            decimal quantity,
+            int quantity,
 
             decimal? price = null,
             TimeInForce? timeInForce = null,
